# internal/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Thoth API",
    "version": "1.0.0",
    "description": "Versioned REST + WebSocket API (v1) for the Thoth local-first knowledge base. Every route lives under `/api/v1/...`; the chat WebSocket upgrades at `/ws/v1`. The server binds to localhost only and requires no authentication; every endpoint assumes a trusted local caller.\n\nError responses use the envelope `{\"error\": \"<message>\"}` \u2014 400 client error, 404 not found, 409 conflict, 500 always the generic `{\"error\": \"internal error\"}` (details go to the server log only).\n\nThe WebSocket chat protocol (`/ws/v1`) is documented in prose in `docs/api.md` (frames and semantics); this spec describes the upgrade only."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8334",
      "description": "Local dev server \u2014 the API reference is served only under serve --dev (the docs page and /swagger.json do not exist in normal serve)"
    }
  ],
...
  "paths": {
...
    "/api/v1/sync/providers": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "List sync providers",
        "description": "The sync provider catalog: built-ins (github, gitlab, aws_s3, local) plus user rows, each with the driver's field descriptors.",
        "operationId": "listSyncProviders",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Create a sync provider",
        "description": "Adds a provider type (a known driver with a custom name/base URL). 400 unknown driver.",
        "operationId": "createSyncProvider",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/syncProvider"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorEnvelope"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "driver"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "driver": {
                    "type": "string",
                    "enum": [
                      "github",
                      "gitlab",
                      "s3",
                      "local"
                    ]
                  },
                  "base_url": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/providers/{id}": {
      "put": {
        "tags": [
          "sync"
        ],
        "summary": "Update a sync provider",
        "description": "Edits name/base_url. Protected providers (local) are 403.",
        "operationId": "updateSyncProvider",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/syncProvider"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorEnvelope"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "base_url": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sync"
        ],
        "summary": "Delete a sync provider",
        "description": "409 while connections reference it; protected providers are 403.",
        "operationId": "deleteSyncProvider",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/connections": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "List connections",
        "description": "Every configured destination with token-free identity and config.",
        "operationId": "listConnections",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Connect a provider",
        "description": "Verifies credentials against the provider and stores the connection. Secrets never echo.",
        "operationId": "connect",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/connection"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorEnvelope"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "provider_id",
                  "name"
                ],
                "properties": {
                  "provider_id": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string"
                  },
                  "config": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/connections/{id}": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Get a connection",
        "description": "One connection with token-free identity and config.",
        "operationId": "getConnection",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorEnvelope"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "sync"
        ],
        "summary": "Update a connection",
        "description": "Edits name/enabled/config. Secret config values with an empty value keep the stored one.",
        "operationId": "updateConnection",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/connection"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorEnvelope"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "enabled": {
                    "type": "boolean"
                  },
                  "config": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sync"
        ],
        "summary": "Disconnect",
        "description": "Removes the connection; the protected local backup is 403.",
        "operationId": "disconnect",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/connections/{id}/targets": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "List sync targets",
        "description": "Selectable destinations for a connected account (repos for git providers; empty otherwise).",
        "operationId": "listTargets",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/connections/{id}/push": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Push the wiki",
        "description": "Syncs the wiki through the connection's driver. Transient failures are retried with backoff; returns ok:false with a sanitized message on failure. Records the outcome (last_synced_at/last_error + push history) on the connection row.",
        "operationId": "pushConnection",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "errorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          }
        }
      },
...
    }
  }
}